### 2020-08-31 7e20453
f/shashank-website -- ReportWriter -- Patient Validation Table generation by adding Parsexml function as well as other functions for that to generate table
