+++ projects/cli/src/utils/ReportWriter.cpp
@@ -319,7 +319,15 @@ struct PatientValidationRow {
   double baseline = 0.0;
   double mean = 0.0;
   double percent_error = 0.0;
+  double expected_value=0.0;
+  double engine_value=0.0 ;
 };
+
+//-------------------------------------------------------------------------------
+/// \brief Takes a list of validation files and results files from biogears and produces
+/// validation tables for our documentation.
+/// \param table_type: char denoting what type of results file should be produced (html, md, or xml)
+//-------------------------------------------------------------------------------
 
 void ReportWriter::gen_patient_tables(TYPE table_type)
 {
@@ -342,6 +350,7 @@ void ReportWriter::gen_patient_tables(TYPE table_type)
     {"Validation-StandardMale","Scenarios/Validation"},
     {"Validation-Overweight","Scenarios/Validation"},
     {"Validation-Carol","Scenarios/Validation"},
+    {"Validation-DefaultTemplateMale","Scenarios/Validation"},
     {"Validation-DefaultTemplateFemale","Scenarios/Validation"},
     {"Validation-Gus","Scenarios/Validation"},
     {"Validation-StandardFemale","Scenarios/Validation"},
@@ -352,19 +361,48 @@ void ReportWriter::gen_patient_tables(TYPE table_type)
     {"Validation-Soldier","Scenarios/Validation"},
      
   };  
+  std::vector<std::string> XMLFile {
+    "states/Cynthia@0s.xml",
+    "states/Joel@0s.xml",
+    "states/Bradycardic@0s.xml",
+    "states/ExtremeFemale@0s.xml",
+    "states/ExtremeMale@0s.xml",
+    "states/Underweight@0s.xml",
+    "states/DefaultMale@0s.xml",
+    "states/Tachycardic@0s.xml",
+    "states/Tachycardic@0s.xml",
+    "states/DefaultFemale@0s.xml",
+    "states/Hassan@0s.xml",
+    "states/Roy@0s.xml",
+    "states/Jeff@0s.xml",
+    "states/ToughGirl@0s.xml",
+    "states/Nathan@0s.xml",
+    "states/StandardMale@0s.xml",
+    "states/Overweight@0s.xml",
+    "states/Carol@0s.xml",
+    "states/DefaultTemplateMale@0s.xml",
+    "states/DefaultTemplateFemale@0s.xml",
+    "states/Gus@0s.xml",
+    "states/StandardFemale@0s.xml",
+    "states/Rick@0s.xml",
+    "states/Jane@0s.xml",
+    "states/ToughGuy@0s.xml",
+    "states/Tristan@0s.xml",
+    "states/Soldier@0s.xml",
+  };  
   bool success = true;
   for (auto& system : SystemTables) {
+      std::vector<std::pair<std::string, PatientValidationRow>> values(34);
+      std::vector<std::pair<std::string,int>> headers;  
       std::vector<std::string>  row_header;
       std::map<std::string, double> mean_table;
       std::string resultsFile =system.second+"/"+system.first + "Results.csv";
-      std::vector<std::pair<std::string, PatientValidationRow>> values(35);
-      std::vector<std::pair<std::string,int>> headers;
       //std::vector<std::string> headers;
       std::fstream result {resultsFile};
       std::string line;
       double row_count=0.0;
+      int num_files=0;
       std::vector<std::string> validation_rows = {
-       "Time(s)", 
        "PatientAge(yr)",
        "PatientWeight(kg)", "PatientHeight(in)",
        "PatientBodyDensity(g/cm^3)","PatientBodyFatFraction",
@@ -390,27 +428,28 @@ void ReportWriter::gen_patient_tables(TYPE table_type)
       while (std::getline(result, line)) {
           if(row_count==0.0){
[…]
