### 2020-08-31 6660012
f/sawhite-removeLogger - Removed logger inheritance from several classes which I feel were superfluous
