+++ projects/howto/BrainInjury/src/HowTo-BrainInjury_main.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 
 #include "HowTo-BrainInjury.h"
-#include "HowToTracker.h"
+
 
 int main( int argc, char* argv[] )
 {
