+++ projects/howto/PulmonaryFunctionTest/src/HowTo-PulmonaryFunctionTest.cpp
@@ -10,13 +10,15 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 **************************************************************************************/
 
-#include "HowToTracker.h"
+
 // Include the various types you will be using in your code
 #include <biogears/cdm/patient/assessments/SEPulmonaryFunctionTest.h>
 #include <biogears/cdm/properties/SEScalarTypes.h>
 #include <biogears/cdm/properties/SEFunctionVolumeVsTime.h>
 #include <biogears/cdm/engine/PhysiologyEngineTrack.h>
 #include <biogears/cdm/compartment/SECompartmentManager.h>
+#include <biogears/engine/BioGearsPhysiologyEngine.h>
+
 
 using namespace biogears;
 //--------------------------------------------------------------------------------------------------
@@ -44,7 +46,7 @@ void HowToPulmonaryFunction()
 	// forced inhalation and exhalation from current tidal volume and engine parameters
 
     // The tracker is responsible for advancing the engine time and outputting the data requests below at each time step
-	HowToTracker tracker(*bg);
+	
 
 	// Create data requests for each value that should be written to the output log as the engine is executing
 	// Physiology System Names are defined on the System Objects 
@@ -60,7 +62,7 @@ void HowToPulmonaryFunction()
 
   bg->GetEngineTrack()->GetDataRequestManager().SetResultsFilename("HowToPulmonaryFunctionTest.csv");
 
-	tracker.AdvanceModelTime(5);
+	bg->AdvanceModelTime(5, TimeUnit::s);
 
 	SEPulmonaryFunctionTest pft(bg->GetLogger());
 	bg->GetPatientAssessment(pft);

[…]
