### 2020-09-04 #2611 @meg-simula
@ajbaird Thanks for submitting BioGears to JOSS. Do you have any suggestions for reviewers?
