### 2020-09-04 #2611 @meg-simula
Also, @ajbaird, could you please take a look at (and fix) the code formatting?
