### 2020-09-07 #2611 @meg-simula
@mirok @fcooper8472 Would you be available to review this submission for JOSS?

Feel free to take a look at our reviewer guidelines here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
