### 2020-09-07 #2645 @whedon
[REVIEW]: BioGears: A C++ library for whole body physiology simulations
**Submitting author:** @ajbaird (<a href="http://orcid.org/0000-0002-4711-3016">Austin Baird</a>)
**Repository:** <a href="https://github.com/BioGearsEngine/core" target ="_blank">https://github.com/BioGearsEngine/core</a>
**Version:** v7.3.2
