### 2020-09-07 #2645 @meg-simula
@ajbaird Could you look at the paper compilation issue reported by @whedon above at your earliest convenience?
