### 2020-09-10 5666923
b/sawhite-cmake -- Added bg-cli abd bg-scenario to install list
