+++ projects/cli/CMakeLists.txt
@@ -109,5 +109,10 @@ if(Boost_FOUND)
    )
 endif()
 ####
+install(TARGETS bg-cli bg-scenario
+     RUNTIME DESTINATION bin
+     LIBRARY DESTINATION ${LIBRARY_INSTALL_DIR}
+     ARCHIVE DESTINATION lib
+)
 message(STATUS "")
 endif()
