### 2020-09-14 #2645 @ajbaird
Hi @meg-simula just pushed an update for the missing DOIs, not totally sure about the compilation error though? Seems to be not able to find the temp/ folder and the cd command is failing? Compiled ok before, maybe it needs to update with the new issue number once it got moved? I'm really not sure, sorry!
