### 2020-09-18 a7a5015
This Patch addresses a bug in master where stablization appears in all csv files regardless of the TrackStabilization value.
