+++ projects/biogears/libBiogears/src/cdm/engine/PhysiologyEngineDynamicStabilization.cpp
@@ -85,19 +85,16 @@ bool PhysiologyEngineDynamicStabilization::Stabilize(PhysiologyEngine& engine, c
   }
   // Execute System initialization time
   PhysiologyEngineTrack* tracker = engine.GetEngineTrack();
-  CDM::enumOnOff::value track = m_TrackingStabilization;
-  if (track && tracker == nullptr) {
-    track = CDM::enumOnOff::Off;
-    Warning("PhysiologyEngineTrack not provided by engine, not tracking data to file");
-  }
-
   bool hasOptionalProperties = false;
 
   // Grab all the convergence properties
   for (PropertyConvergence* pc : properties) {
-    if (pc->IsOptional())
+    if (pc->IsOptional()) {
       hasOptionalProperties = true;
-    tracker->ConnectRequest(pc->GetDataRequest(), pc->GetDataRequestScalar());
+    }
+    if (tracker) {
+      tracker->ConnectRequest(pc->GetDataRequest(), pc->GetDataRequestScalar());
+    }
     if (!pc->GetDataRequestScalar().HasScalar()) {
       ss << "Cannot find convergence property " << pc->GetDataRequest().GetName();
       throw CommonDataModelException(ss.str());
@@ -118,12 +115,8 @@ bool PhysiologyEngineDynamicStabilization::Stabilize(PhysiologyEngine& engine, c
       break;
 
     engine.AdvanceModelTime();
-    if (m_currentTime_s == 0)
-      tracker->SetupRequests();
     stablizationTime_s += dT_s;
     m_currentTime_s += dT_s;
-    if (track)
-      tracker->TrackData(m_currentTime_s);
     if (m_LogProgress) {
       statusTime_s += dT_s;
       if (statusTime_s > statusStep_s) {
@@ -408,7 +401,7 @@ void PhysiologyEngineDynamicStabilization::Unload(CDM::PhysiologyEngineDynamicSt
 //-----------------------------------------------------------------------------
 bool PhysiologyEngineDynamicStabilization::Load(const char* file)
 {
-  return Load(std::string{ file });
+  return Load(std::string { file });
 }
 //-----------------------------------------------------------------------------
 bool PhysiologyEngineDynamicStabilization::Load(const std::string& file)
@@ -458,7 +451,7 @@ const PhysiologyEngineDynamicStabilizationCriteria* PhysiologyEngineDynamicStabi
 //-----------------------------------------------------------------------------
 void PhysiologyEngineDynamicStabilization::RemoveConditionCriteria(const char* name)
 {
-  RemoveConditionCriteria(std::string{ name });
+  RemoveConditionCriteria(std::string { name });
 }
 //-----------------------------------------------------------------------------
 void PhysiologyEngineDynamicStabilization::RemoveConditionCriteria(const std::string& name)
@@ -473,7 +466,7 @@ void PhysiologyEngineDynamicStabilization::RemoveConditionCriteria(const std::st
 //-----------------------------------------------------------------------------
 PhysiologyEngineDynamicStabilizationCriteria* PhysiologyEngineDynamicStabilization::GetConditionCriteria(const char* name) const
 {
-  return GetConditionCriteria(std::string{ name });
+  return GetConditionCriteria(std::string { name });
 }
 //-----------------------------------------------------------------------------
 PhysiologyEngineDynamicStabilizationCriteria* PhysiologyEngineDynamicStabilization::GetConditionCriteria(const std::string& name) const
@@ -578,7 +571,7 @@ void PhysiologyEngineDynamicStabilizationCriteria::SetName(const std::string& na
 //-----------------------------------------------------------------------------
 bool PhysiologyEngineDynamicStabilizationCriteria::HasName() const
 {
[…]
