+++ projects/biogears/libBiogears/src/engine/Controller/BioGearsEngine.cpp
@@ -93,12 +93,12 @@ BioGearsEngine::BioGearsEngine(Logger* logger, const std::string& working_dir)
 }
 //-------------------------------------------------------------------------------
 BioGearsEngine::BioGearsEngine(Logger* logger, const char* working_dir)
-  : BioGearsEngine(logger, std::string{ working_dir })
+  : BioGearsEngine(logger, std::string { working_dir })
 {
 }
 //-------------------------------------------------------------------------------
 BioGearsEngine::BioGearsEngine(const char* logFileName)
-  : BioGearsEngine(std::string{ logFileName })
+  : BioGearsEngine(std::string { logFileName })
 {
 }
 //-------------------------------------------------------------------------------
@@ -146,7 +146,7 @@ PhysiologyEngineTrack* BioGearsEngine::GetEngineTrack()
 //-------------------------------------------------------------------------------
 bool BioGearsEngine::LoadState(const char* file, const SEScalarTime* simTime)
 {
-  return LoadState(std::string{ file }, simTime);
+  return LoadState(std::string { file }, simTime);
 }
 //-------------------------------------------------------------------------------
 bool BioGearsEngine::LoadState(const std::string& file, const SEScalarTime* simTime)
@@ -166,13 +166,13 @@ bool BioGearsEngine::LoadState(const CDM::PhysiologyEngineStateData& state, cons
   auto resultsFile = GetEngineTrack()->GetDataRequestManager().GetResultsFilename();
   std::vector<std::unique_ptr<CDM::DataRequestData>> dataVector;
   for (auto& dr : requests) {
-   dataVector.emplace_back(dr->Unload());
+    dataVector.emplace_back(dr->Unload());
   }
   BioGears::SetUp();
   m_EngineTrack = PhysiologyEngineTrack(*this);
   m_DataTrack = &m_EngineTrack.GetDataTrack();
-  for ( auto& data : dataVector) {
-    m_EngineTrack.GetDataRequestManager().CreateFromBind( *data, *m_Substances); 
+  for (auto& data : dataVector) {
+    m_EngineTrack.GetDataRequestManager().CreateFromBind(*data, *m_Substances);
   }
   m_EngineTrack.GetDataRequestManager().SetResultsFilename(resultsFile);
   m_ss.str("");
@@ -471,7 +471,7 @@ bool BioGearsEngine::LoadState(const CDM::PhysiologyEngineStateData& state, cons
 //-------------------------------------------------------------------------------
 std::unique_ptr<CDM::PhysiologyEngineStateData> BioGearsEngine::SaveState(const char* file)
 {
-  return SaveState(std::string{ file });
+  return SaveState(std::string { file });
 }
 //-------------------------------------------------------------------------------
 std::unique_ptr<CDM::PhysiologyEngineStateData> BioGearsEngine::SaveState(const std::string& file)
@@ -552,7 +552,7 @@ std::unique_ptr<CDM::PhysiologyEngineStateData> BioGearsEngine::SaveState(const 
 //-------------------------------------------------------------------------------
 bool BioGearsEngine::InitializeEngine(const char* patientFile, const std::vector<const SECondition*>* conditions, const PhysiologyEngineConfiguration* config)
 {
-  return InitializeEngine(std::string{ patientFile }, conditions, config);
+  return InitializeEngine(std::string { patientFile }, conditions, config);
 }
 //-------------------------------------------------------------------------------
[…]

[…]
