+++ projects/biogears/libBiogears/src/engine/Systems/Drugs.cpp
@@ -403,82 +403,175 @@ void Drugs::AdministerSubstanceNasal()
   if (nasalDoses.empty())
     return;
 
-  double nasaltime = nasalDoses.at(naloxone)->GetElapsedTime().GetValue(TimeUnit::s);
+  SESubstanceNasalDose* nDose;
+  const SESubstance* nSub;
+  std::vector<const SESubstance*> deactiveSubs;
+  for (auto nd : nasalDoses) {
+    nSub = nd.first;
+    nDose = nd.second;
 
-  double nasaldose = nasalDoses.at(naloxone)->GetDose().GetValue(MassUnit::mg);
+    double nasalTimeElapsed_s = nDose->GetElapsedTime().GetValue(TimeUnit::s);
+    double nasalDose_mg = nDose->GetDose().GetValue(MassUnit::mg);
 
-  //Rate constants in 1/s
-  const double nasalk1 = 0.00001736111;  //translocation rate constant of unreleased substance from the anterior to the posterior section
-  const double nasalk2 = 1000000;  // rate constant of release from drug carrier in anterior section
-  const double nasalk3 = 0.00001736111; // translocation rate constant of released substance from the anterior to the posterior section
-  const double nasalk4 = 0.000173;  // absorption rate constant in anterior section
-  const double nasalk5 = 0.0011575;  // translocation rate constant of unreleased substance from the posterior to the gastrointestinal section
-  const double nasalk6 = 1000000;  // rate constant of release from drug carrier in posterior section
-  const double nasalk7 = 0.0011575; // translocation rate constant of released substance from the posterior to the gastrointestinal section
-  const double nasalk8 = 0.0000260;  // absorption rate constant in posterior section
-  const double nasalk9 = 1000000;  // rate constant of release from drug carrier in gastrointestinal section
-  const double nasalk10 = 0.000000027;  // absorption rate constant in gastrointestinal section
-  const double nasalk11 = 0.00001; // rate constant of released drug degradation in anterior section
-  const double nasalk12 = 0.00001; // rate constant of released drug degradation in posterior section
-  const double nasalk13 = 0.00001; // rate constant of released drug degradation in gastrointestinal section
-  const double nasalk14 = 0.000027777;  // transit rate constant of unreleased drug through gastrointestinal section
-  const double nasalk15 = 0.000027777; // transit rate constant of released drug through gastrointestinal section
-  
-  //Initial Drug Distribution
-  const double nasalA0 = 0; // initial amount of unreleased drug in anterior section 
-  const double nasala0 = 0.6*nasaldose; // initial amount of released drug in anterior section
-  const double nasalP0 = 0; // initial amount of unreleased drug in posterior section 
-  const double nasalp0 = 0.4*nasaldose; // initial amount of released drug in posterior section
-  const double nasalG0 = 0; // initial amount of unreleased drug in gastrointestinal section
-  const double nasalg0 = 0; // initial amount of released drug in gastrointestinal section
+    //Rate constants in 1/s
+    const double nasalk1 = 0.00001736111; //translocation rate constant of unreleased substance from the anterior to the posterior section
[…]
