+++ projects/biogears/libBiogears/src/cdm/patient/actions/SESubstanceNasalDose.cpp
@@ -14,6 +14,7 @@ specific language governing permissions and limitations under the License.
 #include <biogears/schema/cdm/Properties.hxx>
 #include <biogears/cdm/properties/SEScalarTime.h>
 #include <biogears/schema/cdm/PatientActions.hxx>
+#include "biogears/cdm/patient/actions/SESubstanceOralDose.h"
 
 namespace biogears {
 SESubstanceNasalDose::SESubstanceNasalDose(const SESubstance& substance)
@@ -93,6 +94,166 @@ void SESubstanceNasalDose::ToString(std::ostream& str) const
   str << "\n\tDose: ";
   HasDose() ? str << *m_Dose : str << "No Dose Set";
   str << std::flush;
+}
+
+//-----------------------------------------------------------------------------
+//Nasal State methods
+SENasalState::SENasalState(const SESubstance& sub)
+  : m_Substance(&sub)
+{
+  m_TotalNasalDose = nullptr;
+  m_VenaCavaConcentration = nullptr;
+  m_NumUnreleasedMasses = 3; //Hard-coded for current model implementation!
+  m_NumReleasedMasses = 3; //Hard-coded for current model implementation!
+}
+
+SENasalState::~SENasalState()
+{
+  Clear();
+}
+
+void SENasalState::Clear()
+{
+  SAFE_DELETE(m_TotalNasalDose);
+  SAFE_DELETE(m_VenaCavaConcentration);
+  m_UnreleasedDrugMasses.clear();
+  m_ReleasedDrugMasses.clear();
+}
+
+bool SENasalState::Initialize(SEScalarMass& dose)
+{
+  GetTotalNasalDose().Set(dose);
+  GetVenaCavaConcentration().SetValue(0.0, MassPerVolumeUnit::ug_Per_mL);
+  std::vector<double> initUnreleasedMasses(m_NumUnreleasedMasses);
+  std::vector<double> initReleasedMasses(m_NumReleasedMasses);
+  bool unrelSet = SetUnreleasedNasalMasses(initUnreleasedMasses, MassUnit::mg);
+  bool relSet = SetReleasedNasalMasses(initReleasedMasses, MassUnit::mg);
+  return (unrelSet && relSet);
+}
+
+bool SENasalState::Load(const CDM::NasalStateData& in)
+{
+
+  GetTotalNasalDose().Load(in.TotalNasalDose());
+  GetVenaCavaConcentration().Load(in.VenaCavaConcentration());
+  m_UnreleasedDrugMasses.clear();
+  for (auto umData : in.UnreleasedDrugMasses()) {
+    SEScalarMass* unrelMass = new SEScalarMass;
+    unrelMass->Load(umData);
+    m_UnreleasedDrugMasses.push_back(unrelMass);
+  }
+  m_ReleasedDrugMasses.clear();
+  for (auto rmData : in.ReleasedDrugMasses()) {
+    SEScalarMass* relMass = new SEScalarMass;
+    relMass->Load(rmData);
+    m_ReleasedDrugMasses.push_back(relMass);
+  }
+  return true;
+}
+
+CDM::NasalStateData* SENasalState::Unload() const
+{
+  CDM::NasalStateData* data = new CDM::NasalStateData();
+  Unload(*data);
+  return data;
+}
+void SENasalState::Unload(CDM::NasalStateData& data) const
+{
+  data.TotalNasalDose(std::unique_ptr<CDM::ScalarMassData>(m_TotalNasalDose->Unload()));
+  data.VenaCavaConcentration(std::unique_ptr<CDM::ScalarMassPerVolumeData>(m_VenaCavaConcentration->Unload()));
+  for (auto umData : m_UnreleasedDrugMasses) {
[…]
