+++ projects/biogears/libBiogears/src/engine/Systems/Drugs.cpp
@@ -145,6 +145,18 @@ bool Drugs::Load(const CDM::BioGearsDrugSystemData& in)
     otState->Load(otData);
   }
 
+  for (const CDM::NasalStateData& nData : in.NasalStates()) {
+    SESubstance* sub = m_data.GetSubstances().GetSubstance(nData.Substance());
+    if (sub == nullptr) {
+      m_ss << "Unable to find subtance " << nData.Substance();
+      Error(m_ss, "Drugs::Load::NasalAdministration");
+      return false;
+    }
+    SENasalState* nState = new SENasalState(*sub);
+    m_NasalStates[sub] = nState;
+    nState->Load(nData);
+  }
+
   return true;
 }
 CDM::BioGearsDrugSystemData* Drugs::Unload() const
@@ -167,6 +179,11 @@ void Drugs::Unload(CDM::BioGearsDrugSystemData& data) const
   for (auto itr : m_TransmucosalStates) {
     if (itr.second != nullptr)
       data.TransmucosalStates().push_back(std::unique_ptr<CDM::TransmucosalStateData>(itr.second->Unload()));
+  }
+
+  for (auto itr : m_NasalStates) {
+    if (itr.second != nullptr)
+      data.NasalStates().push_back(std::unique_ptr<CDM::NasalStateData>(itr.second->Unload()));
   }
 }
 
@@ -299,7 +316,7 @@ void Drugs::AdministerSubstanceBolus()
       break;
     default:
       /// \error Error: Unavailable Administration Route
-      Error(std::string { "Unavailable Bolus Administration Route for substance " } + b.first->GetName(), "Drugs::AdministerSubstanceBolus");
+      Error(std::string{ "Unavailable Bolus Administration Route for substance " } + b.first->GetName(), "Drugs::AdministerSubstanceBolus");
       completedBolus.push_back(b.first); // Remove it
       continue;
     }
@@ -410,7 +427,23 @@ void Drugs::AdministerSubstanceNasal()
     nSub = nd.first;
     nDose = nd.second;
 
-    double nasalTimeElapsed_s = nDose->GetElapsedTime().GetValue(TimeUnit::s);
+    SENasalState* nState = m_NasalStates[nSub];
+    if (nState == nullptr) {
+      //If it doesn't exist yet, make a new model state for the substance and initialize it
+      nState = new SENasalState(*nSub);
+      if (!nState->Initialize(nDose->GetDose())) {
+        Error("SENasalState::Probable vector length mismatch");
+      }
+      m_NasalStates[nSub] = nState;
+      //Every OT state needs to initialize a GI absorption model state to account for drug that is swallowed.
+      //Clearly we are assuming that there is not already an active pill of the same substance already present in the GI (seems like a safe assumption).
+      //m_data.GetGastrointestinal().NewDrugTransitState(sub);
+      //if (!m_data.GetGastrointestinal().GetDrugTransitState(sub)->Initialize(oDose->GetDose(), oDose->GetAdminRoute())) {
+      //  Error("SEGastrointestinalSystem::SEDrugAbsorptionTransitModelState: Probably vector length mismatch");
+      //}
+    }
+
+    double nasalTimeElapsed_s = m_data.GetTimeStep().GetValue(TimeUnit::s);
     double nasalDose_mg = nDose->GetDose().GetValue(MassUnit::mg);
 
     //Rate constants in 1/s
@@ -431,12 +464,27 @@ void Drugs::AdministerSubstanceNasal()
     const double nasalk15 = 0.000027777; // transit rate constant of released drug through gastrointestinal section
 
     //Initial Drug Distribution
-    const double nasal_anterior_unreleased_initial = 0; // initial amount of unreleased drug in anterior section
[…]
