### 2020-09-29 ee41891
f/sawhite-scenarios -- Minor fixes to scenario files that prevented them from running during validation. Improved driver error output when scenario errors occur
