### 2020-09-30 5ff1a0f
bug/dataRequests: Data request manager duplicate request check is faulty
Data request check for duplicate request only considers Name, which is
ok for Physiology/Patient requests but not for compartment requests
(name field is not unique because you could be looking at different
substances in the same compartment).  Set up hash codes based on name
and (if applicable) compartment and substance fields.  Move check for
duplicates into its own function and call when loading data requests
(because requests were never actually checked when loaded from file).
