+++ projects/biogears/libBiogears/src/cdm/scenario/requests/SEPatientDataRequest.cpp
@@ -44,4 +44,12 @@ void SEPatientDataRequest::Unload(CDM::PatientDataRequestData& data) const
 {
   SEDataRequest::Unload(data);
 }
+
+size_t SEPatientDataRequest::HashCode() const 
+{
+  if (m_Hash == 0) {
+    m_Hash = std::hash<std::string>()(m_Name);
+  }
+  return m_Hash;
+}
 }
