+++ projects/biogears/libBiogears/include/biogears/cdm/scenario/requests/SEEnvironmentDataRequest.h
@@ -26,7 +26,7 @@ public:
   virtual ~SEEnvironmentDataRequest();
 
   virtual void Clear() override; //clear memory
-
+  size_t HashCode() const override;
   virtual bool Load(const CDM::EnvironmentDataRequestData& in);
   virtual CDM::EnvironmentDataRequestData* Unload() const override;
 
