+++ projects/biogears/libBiogears/src/cdm/scenario/requests/SELiquidCompartmentDataRequest.cpp
@@ -43,4 +43,5 @@ void SELiquidCompartmentDataRequest::Unload(CDM::LiquidCompartmentDataRequestDat
 {
   SECompartmentSubstanceDataRequest::Unload(data);
 }
+
 }
