+++ projects/biogears/libBiogears/src/engine/Systems/Drugs.cpp
@@ -416,6 +416,38 @@ void Drugs::AdministerSubstanceNasal()
 {
   SESubstance* naloxone = m_data.GetSubstances().GetSubstance("Naloxone");
 
+  /////////////////////////////////////
+  //const std::map<const SESubstance*, SESubstanceNasalDose*>& nasalDoses = m_data.GetActions().GetPatientActions().GetSubstanceNasalDoses();
+  //if (nasalDoses.empty() && m_NasalStates.empty) //check if no pending actions and not currently active states
+  //  return;
+
+  //SESubstanceNasalDose* nDose;
+  //const SESubstance* nSub;
+  //std::vector<const SESubstance*> deactiveSubs;
+
+  //for (auto nd : nasalDoses) {
+  //  nSub = nd.first;
+  //  nDose = nd.second;
+  //  SENasalState* nState = m_NasalStates[nSub];
+  //  if (m_NasalStates[nSub] = nullptr) {
+  //    //If it doesn't exist yet, make a new model state for the substance and initialize it
+  //    nState = new SENasalState(*nSub);
+  //    if (!nState->Initialize(nDose->GetDose())) {
+  //      Error("SENasalState::Probable vector length mismatch");
+  //    }
+  //    m_NasalStates[nSub] = nState;
+  //  } else {
+  //    drug already active, so this is a multiple dose m_TransmucosalState[sub].GetMouthSolidMass().IncrementValue(od.Dose)
+  //  }
+  //  //Increment mass in first compartment of model (solid mass in mouth) by whatever the new dose is
+  //}
+  //--Clear out oral doses map-- this way the patient oral dose map will be empty and we
+  //  will be able to detect the next time we add a new dose of the same sub
+  //  //Now process transmucosal states (moved out of for loop since we are going to remove doses from patient map
+  //  for (auto state : m_NasalStates)
+  //    Do all the rest of the stuff in the model
+  ////////////////////////////////////
+
   const std::map<const SESubstance*, SESubstanceNasalDose*>& nasalDoses = m_data.GetActions().GetPatientActions().GetSubstanceNasalDoses();
   if (nasalDoses.empty())
     return;
@@ -435,16 +467,11 @@ void Drugs::AdministerSubstanceNasal()
         Error("SENasalState::Probable vector length mismatch");
       }
       m_NasalStates[nSub] = nState;
-      //Every OT state needs to initialize a GI absorption model state to account for drug that is swallowed.
-      //Clearly we are assuming that there is not already an active pill of the same substance already present in the GI (seems like a safe assumption).
-      //m_data.GetGastrointestinal().NewDrugTransitState(sub);
-      //if (!m_data.GetGastrointestinal().GetDrugTransitState(sub)->Initialize(oDose->GetDose(), oDose->GetAdminRoute())) {
-      //  Error("SEGastrointestinalSystem::SEDrugAbsorptionTransitModelState: Probably vector length mismatch");
-      //}
     }
 
     double nasalTimeElapsed_s = m_data.GetTimeStep().GetValue(TimeUnit::s);
-    double nasalDose_mg = nDose->GetDose().GetValue(MassUnit::mg);
+    //double nasalDose_mg = nDose->GetDose().GetValue(MassUnit::mg);
+    double nasalDose_mg = nState->GetTotalNasalDose().GetValue(MassUnit::mg);
 
     //Rate constants in 1/s
     const double nasalk1 = 0.00001736111; //translocation rate constant of unreleased substance from the anterior to the posterior section
[…]
