### 2020-10-05 29bd8e7
bug/dataRequests: Set request hash to 0 when resetting any data
