+++ projects/biogears/libBiogears/src/cdm/scenario/requests/SESubstanceDataRequest.cpp
@@ -80,11 +80,13 @@ const char* SESubstanceDataRequest::GetCompartment() const
 void SESubstanceDataRequest::SetCompartment(const char* name)
 {
   m_Compartment = name;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::SetCompartment(const std::string& name)
 {
   m_Compartment = name;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 bool SESubstanceDataRequest::HasCompartment() const
@@ -105,6 +107,7 @@ const SESubstance* SESubstanceDataRequest::GetSubstance() const
 void SESubstanceDataRequest::SetSubstance(SESubstance* substance)
 {
   m_Substance = substance;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 bool SESubstanceDataRequest::HasSubstance() const
@@ -126,6 +129,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const char* name,
   m_Name = name;
   m_RequestedUnit = "";
   m_Unit = &unit;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::Set(const SESubstance& substance, const std::string& name, const CCompoundUnit& unit)
@@ -137,6 +141,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const std::string
   m_Name = name;
   m_RequestedUnit = "";
   m_Unit = &unit;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::Set(const SESubstance& substance, const char* name, const char* unit)
@@ -148,6 +153,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const char* name,
   m_Name = name;
   m_RequestedUnit = unit;
   m_Unit = nullptr;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::Set(const SESubstance& substance, const std::string& name, const std::string& unit)
@@ -159,6 +165,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const std::string
   m_Name = name;
   m_RequestedUnit = unit;
   m_Unit = nullptr;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::Set(const SESubstance& substance, const char* cmpt, const char* name, const CCompoundUnit& unit)
@@ -170,6 +177,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const char* cmpt,
   m_Name = name;
   m_RequestedUnit = "";
   m_Unit = &unit;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::Set(const SESubstance& substance, const std::string& cmpt, const std::string& name, const CCompoundUnit& unit)
@@ -181,6 +189,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const std::string
   m_Name = name;
   m_RequestedUnit = "";
   m_Unit = &unit;
+  m_Hash = 0;
 }
 //-----------------------------------------------------------------------------
 void SESubstanceDataRequest::Set(const SESubstance& substance, const char* cmpt, const char* name, const char* unit)
@@ -192,6 +201,7 @@ void SESubstanceDataRequest::Set(const SESubstance& substance, const char* cmpt,
   m_Name = name;
   m_RequestedUnit = unit;
   m_Unit = nullptr;
+  m_Hash = 0;
 }
[…]
