+++ share/doc/methodology/BioGears.bib
@@ -8137,3 +8137,10 @@ on Severe Disability and Mortality After Head Injury},
   year={2011},
   publisher={Annual Reviews}
 }
+
+@Book{wilmore2007physiology,
+  Title                    = {Physiology of Sport and Exercise},
+  Author                   = {Wilmore, Jack H., Costill, David, and Kenney, W. Larry},
+  Publisher                = {Human Kinetic Publishers},
+  Year                     = {2007},
+}

[…]
