### 2020-10-06 cd02032
r/abaird-reviewFixes - erasing some comments
