### 2020-10-06 a48633e
r/ntatum-EnergyHRUpdate: fixing review comments
- or statement no longer needed thanks to nervous updates
