+++ projects/biogears/libBiogears/src/engine/Systems/Drugs.cpp
@@ -435,7 +435,7 @@ void Drugs::AdministerSubstanceNasal()
       m_NasalStates[nSub] = nState;
     }
 
-    double nasalDose_mg = nState->GetTotalNasalDose().GetValue(MassUnit::mg);
+    double newNasalDose_mg = nState->GetTotalNasalDose().GetValue(MassUnit::mg);
 
     //Rate constants in 1/s
     const double nasalk1 = 0.00001736111; //translocation rate constant of unreleased substance from the anterior to the posterior section
@@ -459,72 +459,74 @@ void Drugs::AdministerSubstanceNasal()
     std::vector<double> relMass = nState->GetReleasedNasalMasses(MassUnit::mg);
 
     if (0.0 != nDose->GetDose().GetValue(MassUnit::mg)) {
-      relMass[0] += (0.6 * nasalDose_mg); // initial amount of released drug in anterior section
-      relMass[1] += (0.4 * nasalDose_mg); // initial amount of released drug in posterior section
+      relMass[0] += (0.6 * newNasalDose_mg); // initial amount of released drug in anterior section
+      relMass[1] += (0.4 * newNasalDose_mg); // initial amount of released drug in posterior section
     }
-    double nasal_anterior_unreleased_initial = unrelMass[0]; // initial amount of unreleased drug in anterior section
-    double nasal_anterior_released_initial = relMass[0];
-    double nasal_posterior_unreleased_initial = unrelMass[1]; // initial amount of unreleased drug in posterior section
-    double nasal_posterior_released_initial = relMass[1];
-    double nasal_gastro_unreleased_initial = unrelMass[2]; // initial amount of unreleased drug in gastrointestinal section
-    double nasal_gastro_released_initial = relMass[2]; // initial amount of released drug in gastrointestinal section
+    double nasalAnteriorUnreleasedInitial_mg = unrelMass[0]; // initial amount of unreleased drug in anterior section
+    double nasalAnteriorReleasedInitial_mg = relMass[0];
+    double nasalPosteriorUnreleasedInitial_mg = unrelMass[1]; // initial amount of unreleased drug in posterior section
+    double nasalPosteriorReleasedInitial_mg = relMass[1];
+    double nasalGastroUnreleasedInitial_mg = unrelMass[2]; // initial amount of unreleased drug in gastrointestinal section
+    double nasalGastroReleasedInitial_mg = relMass[2]; // initial amount of released drug in gastrointestinal section
 
     //Intermediate Values
-    const double nasala1 = nasalk1 + nasalk2;
-    const double nasalb1 = nasalk5 + nasalk6;
-    const double nasaly1 = nasalk9 + nasalk14;
-    const double nasalo1 = nasalk3 + nasalk4 + nasalk11;
-    const double nasalw1 = nasalk7 + nasalk8 + nasalk12;
-    const double nasale1 = nasalk10 + nasalk13 + nasalk15;
+    const double nasalAlpha = nasalk1 + nasalk2; //alpha
+    const double nasalBeta = nasalk5 + nasalk6; //beta
+    const double nasalGamma = nasalk9 + nasalk14; //gamma
+    const double nasalDelta = nasalk3 + nasalk4 + nasalk11; //delta
[…]
