+++ projects/howto/HemorrhageTreatment/src/HowTo-Hemorrhage_Treatment_main.cpp
@@ -1,0 +1,9 @@
+#include <iostream>
+
+#include "HowTo-Hemorrhage.h"
+
+
+int main( int argc, char* argv[] )
+{
+  HowToHemorrhage();
+}
