+++ projects/howto/HemorrhageTreatment/src/HowTo-Hemorrhage_Treatment.cpp
@@ -1,0 +1,218 @@
+/**************************************************************************************
+Copyright 2015 Applied Research Associates, Inc.
+Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+this file except in compliance with the License. You may obtain a copy of the License
+at:
+http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software distributed under
+the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+**************************************************************************************/
+
+
+
+// Include the various types you will be using in your code
+#include <biogears/cdm/compartment/SECompartmentManager.h>
+#include <biogears/cdm/engine/PhysiologyEngineTrack.h>
+#include <biogears/cdm/patient/actions/SEHemorrhage.h>
+#include <biogears/cdm/patient/actions/SESubstanceCompoundInfusion.h>
+#include <biogears/cdm/properties/SEScalarTypes.h>
+#include <biogears/cdm/substance/SESubstanceCompound.h>
+#include <biogears/cdm/substance/SESubstanceManager.h>
+#include <biogears/cdm/system/physiology/SEBloodChemistrySystem.h>
+#include <biogears/cdm/system/physiology/SECardiovascularSystem.h>
+#include <biogears/cdm/system/physiology/SEEnergySystem.h>
+#include <biogears/engine/BioGearsPhysiologyEngine.h>
+
+using namespace biogears;
+//--------------------------------------------------------------------------------------------------
+/// \brief
+/// Usage for applying a Hemorrhage insult to the patient and also demonstrate how injury code can be used to initiate a hemorrhage if desired
+///
+/// \details
+/// Refer to the SEHemorrhage class
+/// Refer to the SESubstanceManager class
+/// Refer to the SESubstanceIVFluids class for applying an IV to the patient
+//--------------------------------------------------------------------------------------------------
+void ParseMCIS(SEHemorrhage& hem, std::vector<unsigned int>& mcis);
+
+void HowToHemorrhage()
+{
+  // Create the engine and load the patient
+  std::unique_ptr<PhysiologyEngine> bg = CreateBioGearsEngine("HowToHemorrhage.log");
+  bg->GetLogger()->Info("HowToHemorrhage");
+  if (!bg->LoadState("./states/StandardMale@0s.xml")) {
+    bg->GetLogger()->Error("Could not load state, check the error");
+    return;
+  }
+
+  // The tracker is responsible for advancing the engine time and outputting the data requests below at each time step
+  
+
+  // Create data requests for each value that should be written to the output log as the engine is executing
+  // Physiology System Names are defined on the System Objects
+  // defined in the Physiology.xsd file
+  bg->GetEngineTrack()->GetDataRequestManager().CreatePhysiologyDataRequest().Set("HeartRate", FrequencyUnit::Per_min);
[…]
