### 2020-10-13 bd59c45
adding new HowTo for commandline hemorrage treatment application
