### 2020-10-13 3a046b4
Revert "adding new HowTo for commandline hemorrage treatment application"
This reverts commit bd59c455047f88025e6da29a8c6899b54cf15f74.
