### 2020-10-24 06f031c
remove log4cpp
