### 2020-10-25 #2645 @fcooper8472
@arfon  I'm having problems configuring with CMake on Ubuntu 20.04.  I realise there aren't any specific instructions for 20.04, but the problematic dependency is identical to that on 18.04, so perhaps I'm missing something:

I'm following [these instructions](https://github.com/BioGearsEngine/core/wiki/Using-Ubuntu-package-system):

```
installing xsd

wget https://www.codesynthesis.com/download/xsd/4.0/linux-gnu/x86_64/xsd_4.0.0-1_amd64.deb
sudo dpkg -i xsd_4.0.0-1_amd64.deb
```

But get, when running CMake:

```
CMake Warning at cmake/cmake-common_logic.cmake:42 (message):
  The following packages CodeSynthesis were not found.For native compilations
  setting CMAKE_PREFIX_PATH can solve this problemFor cross complilation try
  expanding your CMAKE_FIND_ROOT_PATH
Call Stack (most recent call first):
  projects/biogears/CMakeLists.txt:34 (verify_package)
```

I see that ultimately the script that is responsible for finding XSD is

```
core/cmake/common/FindCodeSynthesis.cmake
```

but I can't see a mechanism to, for instance, explicitly identify paths to search in for XSD.  Any ideas?
