### 2020-10-25 #2645 @fcooper8472
> > @arfon I'm having problems configuring with CMake on Ubuntu 20.04. I realise there aren't any specific instructions for 20.04, but the problematic dependency is identical to that on 18.04, so perhaps I'm missing something:
>
> @ajbaird - I think this question is actually for you :-)

Aha yes, my bad!
