### 2020-10-25 #2645 @ajbaird
Hi @fcooper8472 @arfon ! So I think there are two solutions to the issue you are having, you can opt to use the xsd version provided in the apt-get package manager, I think its called xsdcxx (we haven't tested using this but I think it'll probably work, the brew version works on macOS).

I feel like with our testing, we didn't need to configure cmake prefix paths to point to xsd on linux (when using the .deb package), but you could get the .bz2 file, unzip it and then add a line to the configuration step to set CMAKE_PREFIX_PATH=<path to xsd bin>
