### 2020-10-25 #2645 @fcooper8472
Thanks @ajbaird, got it working by unzipping the bz2 package and specifying `CMAKE_PREFIX_PATH`.  I then get a warning:

```
-- Configuring Biogears
-- Could NOT find Boost: Found unsuitable version "1.71.0", but required is at least "1.73.0" (found /usr/include, )
```

but can't see how to specify the boost location AND the XSD location using `CMAKE_PREFIX_PATH`.

Without boost, though, it seems to have configured and compiled fine.
