### 2020-10-26 ee71920
Integrated pull request from Shashankjain12.  Results look good needs a little more testing and some minor refactoring.
