+++ projects/cli/src/exec/Driver.cpp
@@ -13,15 +13,16 @@
 #include "Driver.h"
 
 #include <functional>
-#include <iostream>
-#include <system_error>
 #include <iomanip>
+#include <iostream>
 #include <mutex>
+#include <system_error>
 
 #if defined(BIOGEARS_SUBPROCESS_SUPPORT)
 #include <boost/asio.hpp>
-#include <boost/process.hpp>
 #include <boost/date_time.hpp>
+#include <boost/process.hpp>
+#include <boost/process/exception.hpp>
 #endif
 
 #include "../utils/Executor.h"
@@ -38,27 +39,24 @@
 #include <biogears/engine/test/BioGearsEngineTest.h>
 #endif
 
+#include "../utils/Config.h"
+#include "../utils/Executor.h"
 #include <biogears/filesystem/path.h>
 #include <biogears/schema/cdm/Scenario.hxx>
 #include <biogears/string/manipulation.h>
 #include <xsd/cxx/tree/exceptions.hxx>
-#include "../utils/Config.h"
-#include "../utils/Executor.h"
-
-
-
 
 #if defined(BIOGEARS_SUBPROCESS_SUPPORT)
-inline std::string  fmt_localtime()
+inline std::string fmt_localtime()
 {
   using namespace boost::posix_time;
   static std::locale loc(std::cout.getloc(),
                          new time_facet("%H:%M"));
 
-  
   std::stringstream ss;
   ss.imbue(loc);
-  ss << boost::posix_time::second_clock::local_time();;
+  ss << boost::posix_time::second_clock::local_time();
+  ;
   return ss.str();
 }
 
@@ -70,9 +68,6 @@ auto cout_loop(boost::process::async_pipe& p, boost::asio::mutable_buffer buf) -
       cout_loop(p, buf);
   });
 }
-
-
-
 
 void cerr_loop(boost::process::async_pipe& p, boost::asio::mutable_buffer buf)
 {
@@ -167,7 +162,7 @@ void Driver::stop_when_empty()
 
     process_s = process_str();
     if (duration < std::chrono::hours(1)) {
-      std::cout << asprintf("\r%sProgress %d/%d; Elapsed Time  %dm%ds\u001b[0K\r", process_s.c_str(), count, _total_work, minutes.count(), duration.count() % 60) << std::flush;
+      std::cout << asprintf("\r%sProgress %d/%d; Elapsed Time  %dm%ds\u001b[0K\r", process_s.c_str(), count, _total_work, minutes.count(), duration.count() % 60);
       std::cout << std::flush;
     } else {
       std::cout << asprintf("\r%sProgress %d/%d; Elapsed Time  %dh%dm%ds\u001b[0K\r", process_s.c_str(), count, _total_work, hours.count(), minutes.count() % 60, duration.count() % 60) << std::flush;
@@ -285,13 +280,13 @@ void Driver::queue_Scenario(Executor exec, bool as_subprocess)
     }
   }
 
-  std::unique_ptr<mil::tatrc::physiology::datamodel::ScenarioData> scenario;
+  std::unique_ptr<CDM::ScenarioData> scenario;
   try {
     xml_schema::flags xml_flags;
     xml_schema::properties xml_properties;
 
     xml_properties.schema_location("uri:/mil/tatrc/physiology/datamodel", "xsd/BioGearsDataModel.xsd");
-    scenario = mil::tatrc::physiology::datamodel::Scenario(ifs, xml_flags, xml_properties);
+    scenario = CDM::Scenario(ifs, xml_flags, xml_properties);
   } catch (std::runtime_error e) {
     std::cout << "Error while processing " << exec.Scenario() << "\n";
     std::cout << e.what() << "\n\n";
[…]
