+++ projects/cli/src/utils/Executor.h
@@ -61,6 +61,7 @@ public:
   EDriver Driver() const;
   EPlotStyle PlotStyle() const;
   bool NoCompare() const;
+  bool TrackStabilization() const;
   std::string Baselines() const;
   std::string Computed() const;
   std::string Group() const;
@@ -73,6 +74,7 @@ public:
   void Driver(EDriver);
   void PlotStyle(EPlotStyle);
   void NoCompare(bool);
+  void TrackStabilization(bool);
   void Baselines(const std::string&);
   void Computed(const std::string&);
   void Group(const std::string&);
@@ -88,10 +90,11 @@ private:
   EDriver driver;
   EPlotStyle plot_style;
   bool no_compare;
+  bool track_stabilization;
   std::string name;
   std::string baselines;
   std::string computed;
-  std::string group;
+  std::string group;                  
   std::string patient;
   std::string state;
   std::string scenario;
