### 2020-10-26 #2645 @ajbaird
Ok, yeah there should just be a check for the version of boost and we rely on a few optional parameters from 1.73, so as long as it compiled ok, I believe you are good to go! We are updating the wiki for your 20.04 build to mention unzipping the bz2 package instead of relying on the .deb package! Thanks for the input.
