### 2020-10-27 df1ac89
Added new subcommand to bg-cli to generate populations of patients based of a new CDM data type PopulationProfile
