+++ projects/cli/src/data/PatientGenerator.cpp
@@ -41,7 +41,7 @@ bool PatientGenerator::save() const
   for (auto& patient : _patients) {
     xml_schema::namespace_infomap info;
     info[""].name = "uri:/mil/tatrc/physiology/datamodel";
-    info[""].schema = "BioGears.xsd";
+    info[""].schema = "BioGearsDataModel.xsd";
 
     try {
       std::ofstream file;
