+++ projects/cli/src/exec/PopulationGenerator.cpp
@@ -1,0 +1,706 @@
+//**********************************************************************************
+//Copyright 2015 Applied Research Associates, Inc.
+//Licensed under the Apache License, Version 2.0 (the "License"); you may not use
+//this file except in compliance with the License.You may obtain a copy of the License
+//at :
+//http://www.apache.org/licenses/LICENSE-2.0
+//Unless required by applicable law or agreed to in writing, software distributed under
+//the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
+//CONDITIONS OF ANY KIND, either express or implied.See the License for the
+//specific language governing permissions and limitations under the License.
+//**************************************************************************************
+
+#include "PopulationGenerator.h"
+
+#ifdef _WIN32
+#include <direct.h>
+#else
+#include <sys/stat.h>
+#endif
+
+#include <algorithm>
+#include <fstream>
+#include <iostream>
+#include <random>
+#include <vector>
+
+#include <biogears/schema/cdm/Patient.hxx>
+#include <biogears/string/manipulation.h>
+
+namespace mil {
+namespace tatrc {
+  namespace physiology {
+    namespace datamodel {
+    }
+  }
+}
+}
+namespace CDM = mil::tatrc::physiology::datamodel;
+
+std::string dateString()
+{
+  std::time_t now = std::chrono::system_clock::to_time_t(std::chrono::system_clock::now());
+
+  std::string s(30, '\0');
+  std::strftime(&s[0], s.size(), "%Y-%m-%dT%H_%M_%S", std::localtime(&now));
+  return s;
+}
+
+namespace biogears {
+std::vector<std::string> generate_boy_names();
+std::vector<std::string> generate_girl_names();
+//-------------------------------------------------------------------------------
+PopulationGenerator::PopulationGenerator(std::vector<std::string> params)
+{
+  if (params.size() == 1) {
+    _runs.emplace_back("PatientTemplate.xml", 10);
+  } else {
+    for (auto i = 0; i < params.size(); ++i) {
+      if (i + 2 <= params.size()) {
+        _runs.emplace_back(params[i], std::stoi(params[i + 1]));
+      }
+      i += 2;
+    }
+  }
+
+  _boy_names = generate_boy_names();
+  _girl_names = generate_girl_names();
+}
+//-------------------------------------------------------------------------------
+void PopulationGenerator::SetGirlNames(std::vector<std::string>&& names)
+{
+  _girl_names = std::move(names);
+}
+//-------------------------------------------------------------------------------
+void PopulationGenerator::SetBoyNames(std::vector<std::string>&& names)
+{
+  _boy_names = std::move(names);
+}
+//-------------------------------------------------------------------------------
+void PopulationGenerator::AddBoyName(std::string name)
+{
+  _boy_names.emplace_back(name);
+}
+//-------------------------------------------------------------------------------
+void PopulationGenerator::AddGirlName(std::string name)
+{
+  _boy_names.emplace_back(name);
+}
+//-------------------------------------------------------------------------------
+void PopulationGenerator::ClearBoyNames()
+{
+  _boy_names.clear();
+}
+//-------------------------------------------------------------------------------
+void PopulationGenerator::ClearGirlNames()
+{
+  _girl_names.clear();
+}
+//-------------------------------------------------------------------------------
+//-------------------------------------------------------------------------------
+//!
+//! \brief Iterates through patientFiles, creates a lambda function for each item, and passes those functions to a thread pool
+//!
+void PopulationGenerator::Generate()
+{
+
+  std::random_device rd {};
+  std::mt19937 gen { rd() };
+  static int total_population_count = 1;
+  std::string unit_str = "";
[…]
