+++ projects/cli/src/data/CompoundGenerator.cpp
@@ -41,7 +41,7 @@ bool CompoundGenerator::save() const
   for (auto& compound : _compounds) {
     xml_schema::namespace_infomap info;
     info[""].name = "uri:/mil/tatrc/physiology/datamodel";
-    info[""].schema = "BioGears.xsd";
+    info[""].schema = "BioGearsDataModel.xsd";
 
     try {
       std::ofstream file;
