+++ projects/cli/src/data/EnvironmentGenerator.cpp
@@ -73,7 +73,7 @@ bool EnvironmentGenerator::save() const
   for (auto& env : _environments) {
     xml_schema::namespace_infomap info;
     info[""].name = "uri:/mil/tatrc/physiology/datamodel";
-    info[""].schema = "BioGears.xsd";
+    info[""].schema = "BioGearsDataModel.xsd";
 
     try {
       std::ofstream file;
