### 2020-10-27 d7556c2
f/sawhite-installer - CMAKE/CPACK - Taught CMake how to use NSIS to create an installer. Will be adding WIX support and theming both before moving on to tarballs and APP distros for macos
