+++ cmake/packaging/README.txt
@@ -1,0 +1,4 @@
+BioGears® is a C++ based, open source, multi-platform (Windows, Mac, and Linux), comprehensive human 
+physiology engine that will drive medical education, research, and training technologies. BioGears enables 
+accurate and consistent physiology simulation across the medical community. The engine can be used as a 
+standalone application or integrated with simulators, sensor interfaces, and models of all fidelities.
