+++ cmake/packaging/CMakeLists.txt
@@ -1,0 +1,60 @@
+include(InstallRequiredSystemLibraries)
+
+set(CPACK_PACKAGE_DESCRIPTION_SUMMARY "BioGears")
+set(CPACK_PACKAGE_VENDOR "Applied Resaerch Associates")
+set(CPACK_RESOURCE_FILE_README "${CMAKE_CURRENT_SOURCE_DIR}/README.txt")
+set(CPACK_RESOURCE_FILE_LICENSE "${CMAKE_CURRENT_SOURCE_DIR}/LICENSE.txt")
+set(CPACK_PACKAGE_HOMEPAGE_URL "http://biogearsengine.com")
+set(CMAKE_PACKAGE_ICON "${CMAKE_CURRENT_SOURCE_DIR}/BioGearsLogo.bmp")
+
+set(CPACK_PACKAGE_INSTALL_DIRECTORY "BioGears/${CMAKE_VERSION_MAJOR}.${CMAKE_VERSION_MINOR}")
+set(CPACK_PACKAGE_VERSION_MAJOR ${${ROOT_PROJECT_NAME}_VERSION_MAJOR} )      
+set(CPACK_PACKAGE_VERSION_MINOR ${${ROOT_PROJECT_NAME}_VERSION_MINOR} )
+set(CPACK_PACKAGE_VERSION_PATCH ${${ROOT_PROJECT_NAME}_VERSION_PATCH} )
+set(CPACK_PACKAGE_VERSION_TWEAK ${${ROOT_PROJECT_NAME}_VERSION_TWEAK} )
+
+
+if(WIN32 AND NOT UNIX)
+  # There is a bug in NSI that does not handle full UNIX paths properly.
+  # Make sure there is at least one set of four backlashes.
+  #set(CPACK_PACKAGE_ICON "${CMAKE_CURRENT_SOURCE_DIR}/BioGearsLogo.bmp")
+  set(CPACK_NSIS_INSTALLED_ICON_NAME "bin\\\\MyExecutable.exe")
+  set(CPACK_NSIS_DISPLAY_NAME "${CPACK_PACKAGE_INSTALL_DIRECTORY} My Famous Project")
+  set(CPACK_NSIS_HELP_LINK "http:\\\\\\\\www.biogearsengine.com")
+  set(CPACK_NSIS_URL_INFO_ABOUT "http:\\\\\\\\www.biogearsengine.com")
+  set(CPACK_NSIS_CONTACT "admin@biogearsengine.com")
+  set(CPACK_NSIS_MODIFY_PATH ON)
+
+  set(CPACK_WIX_UPGRADE_GUID "54B0EF16-257C-4325-A64B-630FFB302F16")
+  set(CPACK_WIX_PRODUCT_ICON "${CMAKE_CURRENT_SOURCE_DIR}/BioGearsLogo.bmp")
+  set(CPACK_WIX_PROGRAM_MENU_FOLDER "BioGears")
+
+  install(DIRECTORY 
+    ${PROJECT_SOURCE_DIR}/share/xsd 
+    DESTINATION runtime/${${ROOT_PROJECT_NAME}_VERSION_TAG}/
+    FILE_PERMISSIONS OWNER_READ GROUP_READ WORLD_READ
+    )
+
+install(DIRECTORY 
+    ${PROJECT_SOURCE_DIR}/share/data/
+    DESTINATION runtime/${${ROOT_PROJECT_NAME}_VERSION_TAG}/
+    FILE_PERMISSIONS OWNER_READ GROUP_READ WORLD_READ
+    FILES_MATCHING 
+    PATTERN "*.xml"
+    PATTERN "*.xlsx"
+    )
+install(DIRECTORY 
[…]
