### 2020-10-30 b167730
Upgraded FindCodeSynthesis to a more generic implementation with less dead code
