+++ cmake/common/FindCodeSynthesis.cmake
@@ -15,9 +15,6 @@
 #   CodeSynthesis_VERSION_STRING - the version of CodeSynthesis found (since CMake 2.8.8)
 
 #=============================================================================
-# Copyright 2006-2009 Kitware, Inc.
-# Copyright 2012 Rolf Eike Beer <eike@sf-mail.de>
-#
 # Distributed under the OSI-approved BSD License (the "License");
 # see accompanying file Copyright.txt for details.
 #
@@ -25,147 +22,112 @@
 # implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 # See the License for more information.
 #=============================================================================
-# (To distribute this file outside of CMake, substitute the full
-#  License text for the above reference.)
 
-#Registers a single XSD File to you cmake PROJECT_LABEL
-#
-# This function assumes that in your project is laid out with a share/xsd folder in PROJECT_SOURCE_DIR
-# schema : filename with out extension to your XSD file to be registered. 
-#          the function expects three functions schema.xsd schema.cfg and schema.xml
-# root_dir : install dir of your public include fodlers example: boost opencv2 
-# component :path to the generated header with respect to root_dir. This will be the location of the generated file in your build_dir
-# resource_path : place where the xml and xsd should be copied to for use in runtime.
+
 
 if(NOT CodeSynthesis_FOUND)
-function(REGISTER_XSD schema)
-    cmake_parse_arguments( "_l" 
-                        "STAGE;INSTALL"
-                        "ROOT;SUBPATH;RESOURCE_FOLDER;TARGETS"
-                        "DEPENDS"
-                        ${ARGN}
-                        )
-
-  set(root_dir ${_l_ROOT} )
-  set(component ${_l_SUBPATH} )
-  set(resource_path ${_l_RESOURCE_FOLDER} )
-  if (NOT _l_TARGETS )
-    set(_l_TARGETS  _REGISTER_XSD_DIR_TARGETS )
-  endif()
-
-  set(CodeSynthesis_FLAGS --output-dir ${CMAKE_CURRENT_BINARY_DIR}/${root_dir}/${component} --options-file ${PROJECT_SOURCE_DIR}/share/xsd/${schema}.cfg ${PROJECT_SOURCE_DIR}/share/xsd/${schema}.xsd)
-
-  if( NOT EXISTS ${CMAKE_CURRENT_BINARY_DIR}/${root_dir}/${component}/${schema}.hxx OR NOT EXISTS ${CMAKE_CURRENT_BINARY_DIR}/${root_dir}/${component}/${schema}.cxx )
-    message(STATUS "Generating ${schema}.hxx and ${schema}.cxx")
-    execute_process( WORKING_DIRECTORY ${PROJECT_SOURCE_DIR}/
-                     COMMAND ${CMAKE_COMMAND} -E make_directory ${CMAKE_CURRENT_BINARY_DIR}/${root_dir}/${component} 
-                     COMMAND ${CMAKE_COMMAND} -E env  LD_LIBRARY_PATH=${${ROOT_PROJECT_NAME}_EXTERNAL}/lib ${CodeSynthesis_EXECUTABLE} cxx-tree ${CodeSynthesis_FLAGS} 
-                     COMMAND ${CMAKE_COMMAND} -E copy ${PROJECT_SOURCE_DIR}/share/xsd/${schema}.xsd ${CMAKE_BINARY_DIR}/${resource_path}/${schema}.xsd   
-                     ERROR_VARIABLE XSD_ERROR
[…]
