### 2020-10-30 4cb0be1
f/sawhite-cmake - CMAKE - Tweaking the target names of STAGE to make it more obvious what it is intended to do.  Also moved howtos to all-howtos so it is the first howto in the list under Examples when a IDE has a Solution Explorer like target tree
