### 2020-11-03 3f98b85
f/ntatum-PlasmaLyteA: md update for compounds
- added a list for all currently supported compounds as well as their
components
