### 2020-11-04 #2645 @fcooper8472
Apologies for the delay in completing my review. I have only one checkbox unticked:

Some general comments on the software/repository:

- The software appears to be well written, tested and documented: in my opinion, well above the threshold for publication in JOSS.
- In the README, under [Contributing](https://github.com/BioGearsEngine/core#contributing), there is a placeholder "Details will be filled in shortly", although enough information is provided that I was happy to check the related box in my review.
- The README contains a table indicating build status on a number of platforms, compilers and architectures. Are those badges "live"? Is the continuous integration hosted somewhere public? Could an interested third party see details of the processes that are behind this table?  There is certainly no requirement for the purpose of this review: I am simply curious as to whether those are just static images, and whether therefore the build is indeed succeeding everywhere.

Some general comments on the paper:

- There is one place with two references ("arterial system by using resistance and compliance elements (ref)(ref)") where the references could be combined.  I believe the syntax would be `[@Ref1; @Ref2]`
- The sentence "Full documentation and validation for every action available to the user is provided through our website link" looks very odd with the link & no punctuation at the end.
