### 2020-11-04 #2645 @ajbaird
Hi @fcooper8472 we use build bot for our CI processes, we have developed it in house using those python libraries to build our cross platform tests, we don't' give access to others to view that splash screen but the badges are generated with buildbot, then uploaded to AWS so the badges are updated nightly. I can def change the general comments this week!

Getting access to our IRC channel is the best way to contribute and we do have a very brief wiki article but I updating the contributing link is something that has been on our radar and needs updating.
