### 2020-11-04 #2645 @fcooper8472
@ajbaird  I think you can just update the paper and then get the bot to rebuild the paper as a new pdf.  Definitely don't need to re-submit I don't think.
