### 2020-11-04 138060e
minor updates to the paper and bibliography after the review phase
