+++ projects/papers/JOSS/paper.bib
@@ -83,5 +83,24 @@
   url = {https://doi.org/10.1109/EMBC.2019.8857686}
 }
 
+@article{willmann2003pk,
+  title={PK-Sim (R): a physiologically based pharmacokinetic'whole-body'model},
+  author={Willmann, Stefan and Lippert, J{\"o}rg and Sevestre, Michael and Solodenko, Juri and Fois, Franco and Schmitt, Walter},
+  journal={Biosilico},
+  volume={4},
+  number={1},
+  pages={121--124},
+  year={2003}
+}
 
+@article{lloyd2004cellml,
+  title={CellML: its future, present and past},
+  author={Lloyd, Catherine M and Halstead, Matt DB and Nielsen, Poul F},
+  journal={Progress in biophysics and molecular biology},
+  volume={85},
+  number={2-3},
+  pages={433--450},
+  year={2004},
+  publisher={Elsevier}
+}
 
