### 2020-11-04 4e5cd19
minor edits, addressing grammer and formating
