+++ projects/papers/JOSS/paper.md
@@ -29,7 +29,7 @@ bibliography: paper.bib
 
 BioGears is an open source, extensible human physiology computational engine that is designed to enhance medical education, research, and training technologies. BioGears is primarily written in C++ and uses an electric circuit analog to characterize the fluid dynamics of the cardiopulmonary system. As medical training requirements become more complex, there is a need to supplement traditional simulators with physiology simulations. To this end, BioGears provides an extensive number of validated injury models and related interventions that may be applied to the simulated patient. In addition, BioGears compiled libraries may be used for computational medical research to construct *in-silico* clinical trials related to patient treatment and outcomes. Variable patient inputs support diversity and specification in a given application. The engine can be used standalone or integrated with simulators, sensor interfaces, and models of all fidelities. The Library, and all associated projects, are published under the Apache 2.0 license and are made available through the public GitHub repository. BioGears aims to lower the barrier to create complex physiological simulations for a variety of uses and requirements.
 
-Physiological models have been used for healthcare simulation for many years but generally, complex models of the system level biology are not implemented and the physiology is preprogrammed by the instructor. The most prominent commercial implementation of similar software is Maestro, developed by Canadian Aviation Electronics, Inc. This product is proprietary and it is not clear to the authors how models are developed, implemented, and validated. Other similar open source projects include: [pk-sim](https://github.com/Open-Systems-Pharmacology/PK-Sim) the pharmacological modeling framework [@willmann2003pk], CellML a generic biological modeling markup language with applications spanning biological applications [lloyd2004cellml], and [Pulse](https://gitlab.kitware.com/physiology/engine) a maintained BioGears fork lacking some of the recent models but instead focusing on Unity VR integration. Each of these either has submodels that are integrated into BioGears (like a pkpd pharmacological model), or is a more generic implementation of biological modeling, like CellML. BioGears is unique in the depth of integrated models, while being free and open-source for the research community.
+Physiological models have been used for healthcare simulation for many years but generally, complex models of the human physiology are not implemented and the patient state is preprogrammed by the instructor. The most prominent commercial implementation of similar software is Maestro, developed by Canadian Aviation Electronics, Inc. This product is proprietary and it is not clear to the authors how models are developed, implemented, and validated. Other similar open source projects include: [pk-sim](https://github.com/Open-Systems-Pharmacology/PK-Sim) the pharmacological modeling framework [@willmann2003pk], CellML a generic biological modeling markup language with applications spanning biological applications [@lloyd2004cellml], and [Pulse](https://gitlab.kitware.com/physiology/engine) a maintained BioGears fork lacking some of the recent models but instead focusing on Unity VR integration. Each of these either has submodels that are integrated into BioGears (like a pkpd pharmacological model), or is a more generic implementation of biological modeling, like CellML. BioGears is unique in the depth of integrated models, while being free and open-source for the research community.
 
 # Statement of need 
 
@@ -37,7 +37,7 @@ The fields of Medical simulation and computational medicine are growing in appli
 
[…]
