### 2020-11-04 #2645 @ajbaird
ok @fcooper8472  just made some changes, I included the bulk of the text changes to address the unchecked box under the summery title. If you think it should have its own header, let me know, I thought it sort of fit there.
