### 2020-11-09 8b5b9d1
readme and documentation for the website clean up and updates
