+++ share/doc/markdown/MainPage.md
@@ -35,44 +35,43 @@ Disclaimer:
 
 - - -
 
-## What's new in ver 6.3 (March 1, 2018)
-The latest deployment includes the following notable updates:
+## What's new in ver 7.3 (January 30, 2020)
+- Custom Compound Infusion 
+- Respiratory Improvements and Tuning
+- Improved website generation targets
+- Updates to CMD_BIO executable for batch validation and scenario execution
+- Website generation python preprocess tool 
+- Transmucosal Fentanyl implementation
+  - oral diffusion through the mucosa layer and a model for GI transit and small intestine absorption
+- Hemorrhage model updates 
+  - Direct model connection to energy and nutrient model 
+- Tourniquet action for use with Hemorrhage scenarios
+- **New drug:** Tranexamic Acid
+  - Validated for use in a hemorrhage scenario to decrease bleed rates 
+- Propofol validation and bug fixes
+- Expansion of the cerebral circuit to a larger, more complex system
+  - Added cerebral auto-regulation, hemorrhage, and updated TBI model
+- **New drugs:** Moxifloxacin and Ertapenem (intra-venous and intra-arterial)
+  - Used in sepsis model, validated for treatment guidelines
+- Validation updates to Fentanyl drug
+  - Updated tissue volumes, updated perfusion limited diffusion method, and updates to PK values  
+- New override functionality
+  - Heart rate, respiration rate, and blood pressure values may now be set while running a conformant engine
+    - Other physiology will change in collaboration with these values to accurately change major physiology during runtime without an injury or action
+- New whole blood model/substance 
+  - Antigens, agglutination model and typed blood substance files for administration during runtime
+- **New drug administration route:** oral tablet 
+  - Antibiotic with validated PK profiles and infection interactions
+  - May be used in sepsis scenarios for management of infection 
+  - New future Tylenol drug for moderate pain management 
+- python plotter util updates to batch plot csv files generated by BioGears
+- New complex How-to files covering burn and sepsis, for developers use
+- Modifications to SE classes to support Unreal Engine integration
+- Updated website generation and layout
+- Patient blood type added as patient parameter
+- General CMAKE updates to the build system 
 
-- General bug fixes, system improvements, and tools/solver improvements
-- Fasciculation patient event flags 
-- Updated sweat methodology (fixes to ions lost in sweat)
-- Updated substance and compound infusion functionality
-  - Added Ringers lactate and updated 
-  - Saline compound ion concentrations corrected
-  - Hardened implementation 
-- MuscleMass new patient data request
-  - Muscle catabolism patient flag
-- Added dehydration condition 
-  - Implemented as scalar 0to1 representing fractional total body water lost
-  - Fluid removed from patient compartments 
-  - Updated patient flag for event and track body weight change (validated)
-  - Added totalbodyfluidVolume as data request
-  - Updated patient weight as a function of condition
-- Added starvation condition
-  - TimeSinceMeal determines how long since the patient's last meal 
-  - Scales internal nutrient storages from validated starvation data
-  - Removed ConsumeMeal condition, now replaced by starvation condition
-  - Validated blood concentrations for ketones, glucose, and amino acids
-  - Updated patient weight as a function of condition
-- Intracellular ion transport
-  - Model uses membrane potential  (see @ref TissueMethodology for details)
-  - Michaelis coefficient could support more ion regulation in the future
[…]
