### 2020-11-15 #2645 @MiroK
@ajbaird I apologize for late response. From my point of view all looks/works nicely; I was able to install and run the code without issues on my Ubuntu 18.04. The only comments I have are regarding the article proof:

- In the sentence on page 2. "compliance elements(Otto, 1899; Westerhof et al., 2009)" I guess there should be a space
- " The software architecture of BioGears is implemented in three layers of abstraction to encourage easy integration and provide a robust application programming interface (API), Figure 1". Should this be, "see Figure 1." or similar?
