### 2020-12-03 #2645 @meg-simula
@ajbaird

Please address the following minor points in the paper:
- [ ] Add space before reference to (Otto...)
- [ ] This phrasing is unclear, please rephrase: "This work has extended the engine to support ..."
- [ ] "New dugs" should probably be "New drugs" - please spell-check
- [ ] Missing punctuation after reference to Figure 1
- [ ] Should HowToFaciculation() be HowToFasciculation() ?

After these points have been addressed, could you:
