+++ projects/papers/JOSS/paper.bib
@@ -101,6 +101,7 @@
   number={2-3},
   pages={433--450},
   year={2004},
+  doi = {10.1016/j.pbiomolbio.2004.01.004},
   publisher={Elsevier}
 }
 
