### 2020-12-03 #2645 @ajbaird
Hi @meg-simula ! I think I got to all the edits you listed here with a few other minor things.

- Software version is v7.3.2 https://github.com/BioGearsEngine/core/releases/tag/7.3.2
-  release is up on Zenodo, DOI is here 10.5281/zenodo.4304604, Author list and title should be the same as this paper.

I think that is everything, let me know if I got to it, sorry if I missed something.

Thanks!

Austin
