### 2020-12-14 #2645 @arfon
@ajbaird - could you please add a little more information to the affiliation string? For example, city and country name.
