### 2020-12-14 #2645 @arfon
@fcooper8472, @MiroK - many thanks for your reviews here and to @meg-simula for editing this submission. JOSS relies upon the volunteer efforts of folks like yourself - we simply couldn't do it without you ✨
